requirements.txt
telebot==0.0.5
pyTelegramBotAPI>=4.12.0
web3>=6.0.0
python-dotenv>=1.0.0
py-solc-x>=1.1.1
requests>=2.28.2
python-telegram-bot>=13.15 